Errors are fixed by Beka

+++ app/src/main/java/ru/startandroid/weather/response/Main.java
@@ -1,13 +1,13 @@
 package ru.startandroid.weather.response;
 
 public class Main {
-    public int temp;
+    public float temp;
 
-    public int getTemp() {
+    public float getTemp() {
         return temp;
     }
 
-    public void setTemp(int temp) {
+    public void setTemp(float temp) {
         this.temp = temp;
     }
